requirement.txt
asgiref==3.5.2
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
